Fix code standard errors
